[DEBUG] MAC BUILD ERROR IN LOG

+++ etk/log.cpp
@@ -425,16 +425,16 @@ void etk::log::logChar(int32_t _id, int32_t _level, int32_t _ligne, const char* 
 			if (startPos != nullptr) {
 				if (stopPos != nullptr) {
 					if(stopPos < startPos) {
-						snprintf(tmpPointer, std::min(1024, stopPos-_funcName), "%s", _funcName);
+						snprintf(tmpPointer, std::min(1024, int32_t(stopPos-_funcName)), "%s", _funcName);
 					} else {
-						snprintf(tmpPointer, std::min(1024, stopPos-startPos), "%s", startPos+1);
+						snprintf(tmpPointer, std::min(1024, int32_t(stopPos-startPos)), "%s", startPos+1);
 					}
 				} else {
 					snprintf(tmpPointer, 1024, "%s", startPos);
 				}
 			} else {
 				if (stopPos != nullptr) {
-					snprintf(tmpPointer, std::min(1024, stopPos-_funcName), "%s", _funcName);
+					snprintf(tmpPointer, std::min(1024, int32_t(stopPos-_funcName)), "%s", _funcName);
 				} else {
 					snprintf(tmpPointer, 1024, "%s", _funcName);
 				}
